pull init out of the manager
